requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.1
bs4==0.0.1
contourpy==1.0.7
cycler==0.11.0
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
fonttools==4.39.3
kiwisolver==1.4.4
lxml==4.9.2
matplotlib==3.7.1
mysqlclient==2.1.1
numpy==1.24.2
packaging==23.0
pandas==2.0.0
Pillow==9.5.0
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
tzdata==2023.3
xmltodict==0.13.0